Implement delivery of notifications on reports creation (#67)
